% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_thresholds}
\alias{apply_thresholds}
\title{Choose the correct intensity thresholds}
\usage{
apply_thresholds(intensity, ages, device, wear_location)
}
\arguments{
\item{intensity}{The thresholds to return. Must be one of 'light', 'mod', or
'vig'}

\item{ages}{The age group (i.e., children or adults or both)}

\item{device}{The device (i.e., actigraph or geneactiv or both)}

\item{wear_location}{The wear location of the device (i.e., 'wrist' or 'hip')}
}
\value{
vector of threshold cut-points
}
\description{
A helper function for choosing intensity thresholds. Not intended to be
invoked directly by the user.
}
\seealso{
Other helper functions (internal): 
\code{\link{check_outfiles}()},
\code{\link{translate_studyname}()}
}
\concept{helper functions (internal)}
