% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{assemble_datasheet}
\alias{assemble_datasheet}
\title{Assemble Datasheet}
\usage{
assemble_datasheet(folder, verbose = TRUE)
}
\arguments{
\item{folder}{The folder that \code{\link[=collate_outputs]{collate_outputs()}} (or \code{\link[=reprocess]{reprocess()}}) wrote to
during reprocessing.}

\item{verbose}{Should a success messages be displayed? \code{TRUE} by default.}
}
\value{
Nothing is returned. A new datasheet is saved to the folder.
}
\description{
Find the outputs of reprocessing and assemble them into a datasheet.
}
\examples{
\dontrun{
outfolder <- "C:\\\\Documents\\\\My Study\\\\Output\\\\sleepIPD_output"
assemble_datasheet(outfolder)
}
}
