% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_meta}
\alias{build_meta}
\title{Create Pre-filled Metadata Template}
\usage{
build_meta(outputdir, studyname, overwrite = FALSE, wear_location = NULL)
}
\arguments{
\item{outputdir}{The directory where the output files were created.}

\item{studyname}{The name of the study. Use the first author's surname where
there is no study name.}

\item{overwrite}{If an existing metadata template is found, should it be
replaced? Default is \code{FALSE}.}

\item{wear_location}{Optionally pre-fill the wear location field of the
metadata. If there are multiple wear locations, this is ignored.}
}
\description{
Create Pre-filled Metadata Template
}
\seealso{
Other helper functions: 
\code{\link{collate_outputs}()},
\code{\link{find_problem_files}()},
\code{\link{translate_filepath}()}
}
\concept{helper functions}
