% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_outfiles}
\alias{check_outfiles}
\title{Check Number of Files Matches Expectations}
\usage{
check_outfiles(part, expect, found, verbose)
}
\arguments{
\item{part}{The GGIR part the files relate to (as a string).}

\item{expect}{The number of files expected to be found.}

\item{found}{The number of files actually found.}

\item{verbose}{Should a success message be displayed?}
}
\description{
An internal function. Performs a check for output files to confirm the
correct number are found.
}
\seealso{
Other helper functions (internal): 
\code{\link{apply_thresholds}()},
\code{\link{translate_studyname}()}
}
\concept{helper functions (internal)}
