% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collate_outputs}
\alias{collate_outputs}
\title{Collate GGIR Outputs}
\usage{
collate_outputs(outputdir, studyname, verbose)
}
\arguments{
\item{outputdir}{the outputdir that was provided in the processing call}

\item{studyname}{the name of the study. Use the first author's surname if
there is no study name.}

\item{verbose}{If success messages should be displayed}
}
\description{
This function simply finds the outputs from GGIR that are required for the
project, and copies them to a new folder. If there are multiple copies of
part 5 outputs, they are collapsed to a single csv file.
}
\seealso{
Other helper functions: 
\code{\link{build_meta}()},
\code{\link{translate_filepath}()}
}
\concept{helper functions}
