% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{reprocess}
\alias{reprocess}
\title{Reprocess Accelerometer Data and Generate Metadata}
\usage{
reprocess(
  datadir,
  outputdir,
  studyname,
  ages = c("children", "adults"),
  device = c("geneactiv", "actigraph"),
  wear_location = c("wrist", "hip"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{datadir}{The directory with the files to process or a vector of
file paths to process (as per \code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}).}

\item{outputdir}{the directory to store the output (as per
\code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}).}

\item{studyname}{The name of the study. This is required to name the
metadata.}

\item{ages}{The age of the sample. Can be either \code{"children"} (for those 18
or less), \code{"adults"} (for >18) or both (i.e., \code{c("children", "adults")}). If
you do not specify, both are used.}

\item{device}{The device used in the study. Can be either \code{"geneactiv"},
\code{"actigraph"} or (in the unlikely event two devices were used) both (i.e.,
\code{c("geneactiv", "actigraph")}). If you do not specify, both are used.}

\item{wear_location}{The location the device was warn. Can be either
\code{"wrist"}, \code{"hip"} or (in the unlikely event multiple locations were used)
both (i.e., \code{c("wrist","hip")}). If you do not specify, both are used.}

\item{overwrite}{If the existing output (including metadata) should be
replaced. \code{FALSE} by default.}

\item{verbose}{Determines how much output the function returns. Useful for
understanding what the package is doing. \code{TRUE} by default.}

\item{...}{Pass additional parameters to \code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}. Refer to the
\code{GGIR} manual for details.}
}
\description{
This is the primary wrapper function for sleepIPD. In most cases, this should
be the only function that is needed. \code{reprocess} collects data about your
study, and then calls the functions to reprocess the data, collate the
results, and generate the metadata.
}
\details{
The purpose of sleepIPD is to standardise the processing of the
accelerometery files. As such, most of the arguments come preset and should
not be modified. However, options which do not affect the output (e.g., the
decision to process in parallel or not) are available via \code{...}. See
\code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}} for details of these arguments.

You can specify parameters of your study by supplying \code{ages}, \code{device},
and/or \code{wear_location} (see Arguments for details). Not supplying these will
cause sleepIPD to calculate every combination of thresholds (resulting in
\eqn{8^3 \times 2 = 1024} output files). So, supplying them should slightly
reduce processing time.

You should also supply a \code{studyname} argument. This is required by GGIR if
your \code{datadir} is a vector of files, but is additionally used by sleepIPD to
name the output files. If your study did not have a name, the last name of
the first author is sufficent (if you are supplying multiple studies, name
each one with the year the data were collected; e.g.,
\code{studyname = 'Hartwig 2021'}).
}
\examples{
\dontrun{
datadir <- "C:\\\\Documents\\\\My Study\\\\Data"
outputdir <- "C:\\\\Documents\\\\My Study\\\\Output"
studyname <- "Hartwig 2021"
ages <- "children"
device <- "geneactiv"
wear_location <- "wrist"

reprocess(datadir = datadir,
          outputdir = outputdir,
          studyname = studyname,
          ages = ages,
          device = device,
          wear_location = wear_location)
}
}
\seealso{
Other processing functions: 
\code{\link{process_files}()}
}
\concept{processing functions}
