% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepIPD-package.R
\docType{package}
\name{sleepIPD-package}
\alias{sleepIPD}
\alias{sleepIPD-package}
\title{sleepIPD: Opinionated Accelerometer Reprocessing for the SleepIPD Study}
\description{
A wrapper around {GGIR} to provide standardised reprocessing of accelerometer files. It also provides a set of helper scripts to make generating the metadata simpler.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://motivation-and-behaviour.github.io/sleepIPD}
  \item \url{https://github.com/Motivation-and-Behaviour/sleepIPD}
  \item Report bugs at \url{https://github.com/Motivation-and-Behaviour/sleepIPD/issues}
}

}
\author{
\strong{Maintainer}: Taren Sanders \email{taren.sanders@acu.edu.au} (\href{https://orcid.org/0000-0002-4504-6008}{ORCID})

}
\keyword{internal}
