% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translate_studyname}
\alias{translate_studyname}
\title{Remove White Space from a Study Name}
\usage{
translate_studyname(studyname)
}
\arguments{
\item{studyname}{Name of the study}
}
\value{
Study name with whitespace replaced with \code{-}
}
\description{
Remove White Space from a Study Name
}
\seealso{
Other helper functions (internal): 
\code{\link{apply_thresholds}()},
\code{\link{check_outfiles}()}
}
\concept{helper functions (internal)}
