% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_GGIR}
\alias{validate_GGIR}
\title{Check GGIR Version}
\usage{
validate_GGIR(verbose = TRUE)
}
\arguments{
\item{verbose}{Should a success message be displayed? \code{TRUE} by default.}
}
\description{
Performs a check of the installed version of GGIR. The user is invited to
install the correct version if it cannot be located.
}
\concept{validation functions}
