% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_files}
\alias{process_files}
\title{Process files to SleepIPD specs}
\usage{
process_files(
  datadir,
  outputdir,
  studyname = "NA",
  ages = c("children", "adults"),
  device = c("geneactiv", "actigraph"),
  wear_location = "wrist",
  overwrite = FALSE,
  verbose = TRUE,
  redo = FALSE,
  ...
)
}
\arguments{
\item{datadir}{The directory with the files to process or a vector of
file paths to process (as per \code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}).}

\item{outputdir}{the directory to store the output (as per
\code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}).}

\item{studyname}{The name of the study. Only used if \code{datadir} is a vector of
files.}

\item{ages}{The age of the sample. Can be either \code{"children"} (for those 18
or less), \code{"adults"} (for >18) or both (i.e., \code{c("children", "adults")}). If
you do not specify, both are used.}

\item{device}{The device used in the study. Can be either \code{"geneactiv"},
\code{"actigraph"} or (in the unlikely event two devices were used) both (i.e.,
\code{c("geneactiv", "actigraph")}). If you do not specify, both are used.}

\item{wear_location}{The location the device was warn. Can be either
\code{"wrist"} (the default) or \code{"hip"} If the study has mixed wear locations,
you need to run the function twice, once for each location.}

\item{overwrite}{If the existing output should be replaced. \code{FALSE} by
default.}

\item{verbose}{Determines how much output the function returns. Useful for
understanding what the package is doing. \code{TRUE} by default.}

\item{redo}{For advanced usecases. If \code{TRUE}, the function will redo parts
2 through 5, assuming that part 1 has already been completed.}

\item{...}{Pass additional parameters to \code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}}. Refer to the
\code{GGIR} manual for details.}
}
\description{
\code{process_files} collects information about your study sample and reprocesses
the files accordingly.
}
\details{
The purpose of this function is to standardise the processing of the
accelerometery files. As such, most of the arguments come preset and should
not be modified. However, options which do not affect the output (e.g., the
decision to process in parallel or not) are available via \code{...}. See
\code{\link[GGIR:g.shell.GGIR]{GGIR::g.shell.GGIR()}} for details of these arguments.

You can specify parameters of your study by supplying \code{ages}, \code{device},
and/or \code{wear_location} (see Arguments for details). Not supplying these will
cause sleepIPD to calculate every combination of thresholds (resulting in
\eqn{8^3 \times 2 = 1024} output files). So, supplying them should slightly
reduce processing time.
}
\seealso{
Other processing functions: 
\code{\link{reprocess}()}
}
\concept{processing functions}
