% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translate_filepath}
\alias{translate_filepath}
\title{Convert a File or Folder Path}
\usage{
translate_filepath(path)
}
\arguments{
\item{path}{The file or folder path to convert.}
}
\value{
The same file path with \code{\\\\} replaced with \code{/}
}
\description{
GGIR uses some code which does not play nice with some file paths on Windows
(e.g., those produced by \code{tempdir()}). This function converts between a
file path to the format that works with GGIR.
}
\examples{
example_path <- file.path("C:", "Documents", "Study Files", fsep = "\\\\")
translate_filepath(example_path)
}
